+++ packages/adapter/base/signer.ts
@@ -1,4 +1,4 @@
-import type { NoteFilterTypes } from '@demox-labs/miden-sdk';
+import type { NoteFilterTypes } from '@miden-sdk/miden-sdk';
 import type { WalletAdapter, WalletAdapterProps } from './adapter';
 import { BaseWalletAdapter } from './adapter';
 import {