+++ packages/adapter/react/WalletProvider.tsx
@@ -12,13 +12,13 @@ import {
   WalletNotReadyError,
   WalletNotSelectedError,
   WalletReadyState,
-} from '@demox-labs/miden-wallet-adapter-base';
+} from '@miden-sdk/miden-wallet-adapter-base';
 import type { FC, ReactNode } from 'react';
 import { useCallback, useEffect, useMemo, useRef, useState } from 'react';
 import { useLocalStorage } from './useLocalStorage';
 import type { Wallet } from './useWallet';
 import { WalletContext } from './useWallet';
-import type { NoteFilterTypes } from '@demox-labs/miden-sdk';
+import type { NoteFilterTypes } from '@miden-sdk/miden-sdk';
 
 export interface WalletProviderProps {
   children: ReactNode;