+++ packages/adapter/reactui/src/WalletConnectButton.tsx
@@ -1,10 +1,10 @@
 import type { FC, MouseEventHandler } from 'react';
 import { useCallback, useMemo } from 'react';
-import { useWallet } from '@demox-labs/miden-wallet-adapter-react';
+import { useWallet } from '@miden-sdk/miden-wallet-adapter-react';
 import type { ButtonProps } from './Button';
 import { Button } from './Button';
 import { WalletIcon } from './WalletIcon';
-import { AllowedPrivateData, PrivateDataPermission, WalletAdapterNetwork } from '@demox-labs/miden-wallet-adapter-base';
+import { AllowedPrivateData, PrivateDataPermission, WalletAdapterNetwork } from '@miden-sdk/miden-wallet-adapter-base';
 
 export const WalletConnectButton: FC<ButtonProps> = ({
   children,