+++ packages/adapter/reactui/src/WalletDisconnectButton.tsx
@@ -1,6 +1,6 @@
 import type { FC, MouseEventHandler } from 'react';
 import { useCallback, useMemo } from 'react';
-import { useWallet } from '@demox-labs/miden-wallet-adapter-react';
+import { useWallet } from '@miden-sdk/miden-wallet-adapter-react';
 import type { ButtonProps } from './Button';
 import { Button } from './Button';
 import { WalletIcon } from './WalletIcon';