+++ packages/adapter/reactui/src/WalletMultiButton.tsx
@@ -6,7 +6,7 @@ import {
   useRef,
   useState,
 } from 'react';
-import { useWallet } from '@demox-labs/miden-wallet-adapter-react';
+import { useWallet } from '@miden-sdk/miden-wallet-adapter-react';
 import type { ButtonProps } from './Button';
 import { Button } from './Button';
 import { useWalletModal } from './useWalletModal';